+++ R/ipfN2marginals.R
@@ -1,8 +1,9 @@
 #' Iterative proportional fitting of an abundance table to Hill-N2 marginals
 #' 
-#' Function for preprocessing/transforming an abundance table
+#' Function for pre-processing/transforming an abundance table
 #' by iterative proportional fitting,  
-#' so that the transformed table has marginals \code{N2} or \code{N2(N-N2)}
+#' so that the transformed table has marginals
+#' proportional to \code{N2} or \code{N2(1-N2/N)}
 #' with \code{N} the number of elements in the margin.
 #' Hill-N2 is the effective number of species. It is of intrinsic interest in
 #' weighted averaging (CWM and SNC) as their variance is approximately 
@@ -17,7 +18,7 @@
 #' the N2-marginals of the initial table, but the N2-marginals of the returned 
 #' matrix may not be equal to their marginal sum.
 #' If \code{updateN2 = TRUE} and \code{N2N_N2_species=TRUE} (the default), 
-#' the column marginals are \code{N2(N-N2)/2} with \code{N} the number of sites. 
+#' the column marginals are \code{N2(N-N2)/N} with \code{N} the number of sites.
 #' The row sums are then proportional to, what we term, the effective number of
 #' informative species.
 #' If \code{N2N_N2_species = FALSE},								
@@ -26,11 +27,16 @@
 #' the return value of \code{ipf2N2} and \code{const} a constant.
 #' If converged, N2 row marginals are equal to the row sums, \emph{i.e.} 
 #' \code{rowSums(Y2) = approx. N2sites(Y2)}.
-#' @param N2N_N2_species Set marginals proportional to \code{N2(N-N2)} Default 
+#' @param N2N_N2_species Set marginals proportional to \code{N2(1-N2/N)} Default
 #' \code{TRUE}.
 #' @param N2N_N2_sites Default \code{FALSE}. Do not change.
 #' @param max_iter maximum number of iterative proportional fitting (ipf) 
 #' iterations.
+#' If \code{max_iter == 0}, the columns are divided by their 
+#' informativeness (\code{N2}) or \code{N2(1-N2/N)}) without
+#' further pre-processing. The row sums are sums of 
+#' informativeness instead of effective number of informative species.
+#'
 #' @param crit stopping criterion.
 #' 
 #' @return a matrix of the same order as the input \code{Y},
@@ -58,7 +64,7 @@
 #' 
 #' @export
 ipf2N2 <- function(Y, 
-                   max_iter = 100, 
+                   max_iter = 1000, 
                    updateN2 = TRUE, 
                    N2N_N2_species = TRUE, 
                    N2N_N2_sites = FALSE, 
@@ -78,32 +84,43 @@ ipf2N2 <- function(Y,
     stop("Some species are absent in every site.\n", 
          names(K)[which(K == 0)], "\n")
   }
-  N2spp <- fN2N_N2(Y, 2, N2N_N2 = N2N_N2_species)
-  N2sites <- fN2N_N2(Y, 1, N2N_N2 = N2N_N2_sites)
+  N2spp <- N2spp0 <- fN2N_N2(Y, 2, N2N_N2 = N2N_N2_species)
+  N2sites <- N2sites0 <- fN2N_N2(Y, 1, N2N_N2 = N2N_N2_sites)
   N <- nrow(Y)
   tot <- sum(R)
   iter <- 0
   # the criterion is on the weights made relative (what counts in CA)
   while (iter == 0 || 
          (max(N2sites / sum(N2sites) - R / sum(R)) > crit / N && iter < max_iter)) { 
-    iter <- iter + 1
     Y <- Y * (N2sites / R)
     K <- colSums(Y)
+    K[K < .Machine$double.eps] <- .Machine$double.eps		 
     if (updateN2) N2spp <- fN2N_N2(Y, 2, N2N_N2 = N2N_N2_species)
     Y <- Y %*% diag(N2spp / K)
+    if (max_iter == 0) break
+    iter <- iter + 1
     R <- rowSums(Y)
+    R[R< .Machine$double.eps] <- .Machine$double.eps												
     if (updateN2) N2sites <- fN2N_N2(Y, 1, N2N_N2 = N2N_N2_sites)
   }
-  if (iter == max_iter) {
+  if (iter == max_iter && max_iter > 0) {
     warning(paste0("No convergence in ", max_iter, " iterations."),"\n")
   }
+  Y[Y < 0] <- 0 # tiny non-negative values do occur in practice														
   Y <- Y * (tot / sum(Y))
+  attr(Y, which = "N2species_original") <- N2spp0
+  attr(Y, which = "N2sites_original") <- N2sites0
   attr(Y, which = "N2species") <- N2spp
   attr(Y, which = "N2sites") <- N2sites
   attr(Y, which = "R/N2") <- (R / N2sites) / (sum(R) / sum(N2sites))
   attr(Y, which = "iter") <- iter
   colnames(Y) <- colnams
   rownames(Y) <- rownams
+  R <- rowSums(Y)
+  if (fN2(R) / length(R) < 0.5) {
+    message("Warning: unbalanced site totals in return value:",
+            "N2 of row sums less than halve the number of rows.\n")
+  }
   return(Y)
 }
 
@@ -141,6 +158,7 @@ fN2N_N2 <- function(Y,
   N2 <- apply(X = Y, MARGIN = margin, FUN = fN2)
   # the reason for the 2 is to strech the number to N/2 when N2 = N/2
   margin1 <- if (margin == 1) 2 else 1
-  if (N2N_N2) N2 <- 2 * N2 * (1 - N2 / dim(Y)[margin1])
+  if (N2N_N2) N2 <- N2 * (1 - N2 / dim(Y)[margin1])
+  N2[is.na(N2)] <- 0
   return(N2)
 }